etl-weatherstack/requirements.txt
requests==2.31.0
pandas==2.1.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.20
python-dotenv==1.0.0
pydantic==2.4.0
schedule==1.2.0
numpy==1.24.3
